refactor : 플레이리스트 응답값 수정
